Remove part of the piping function
Normally if piping is set to a user that user doesn't have to mention the bot to get the message to send. I removed this feature but kept piping because I have Cleverbot talking to Omeglebot, so I need omeglebot to mention Cleverbot. Cleverbot always responds with a mention, so that would get sent to the user. That is not desirable in my use case.